Add user's name at top of screen
