+++ src/domUpdates.js
@@ -1,5 +1,7 @@
 let cardArea = document.querySelector('.all-cards');
 
+
+
 export const domUpdate = {
     populateCards(recipes) {
         cardArea.innerHTML = '';
@@ -25,14 +27,46 @@ export const domUpdate = {
             src='${recipe.image}' alt='click to view recipe for ${recipe.name}'>
         </div>`)
     });
-    getFavorites();
     },
 
-    getFavorites() {
-        if (user.favoriteRecipes.length) {
-            user.favoriteRecipes.forEach(recipe => {
-                document.querySelector(`.favorite${recipe.id}`).classList.add('favorite-active')
+    displayDirections(event) {
+        apiData()
+        .then(data => {
+            let newRecipeInfo = reciperepo.recipes.find(recipe => {
+                if (recipe.id === Number(event.target.id)) {
+                return recipe;
+            }})
+            let recipeObject = new Recipe(newRecipeInfo, data.ingredientsData);
+            let cost = recipeObject.calculateCost()
+            let costInDollars = (cost / 100).toFixed(2)
+            cardArea.classList.add('all');
+            cardArea.innerHTML = `<h3>${recipeObject.name}</h3>
+            <p class='all-recipe-info'>
+            <strong>It will cost: </strong><span class='cost recipe-info'>
+            $${costInDollars}</span><br><br>
+            <strong>You will need: </strong><span class='ingredients recipe-info'></span>
+            <strong>Instructions: </strong><ol><span class='instructions recipe-info'>
+            </span></ol>
+            </p>`;
+            let ingredientsSpan = document.querySelector('.ingredients');
+            let instructionsSpan = document.querySelector('.instructions');
+            recipeObject.ingredients.forEach(ingredient => {
+            ingredientsSpan.insertAdjacentHTML('afterbegin', `<ul><li>
+            ${ingredient.quantity.amount.toFixed(2)} ${ingredient.quantity.unit}
+            ${ingredient.name}</li></ul>
+            `)
             })
-        } else return
+            recipeObject.instructions.forEach(instruction => {
+                instructionsSpan.insertAdjacentHTML('beforebegin', `<li>
+                ${instruction.instruction}</li>
+                `)
+            })
+        })
+    },
+
+    greetUser(user) {
+        const userName = document.querySelector('.user-name');
+        userName.innerHTML =
+        user.name.split(' ')[0] + ' ' + user.name.split(' ')[1][0];
     }
 };
